backend/requirements.txt
flask
jinja2==3.0.3
requests
cors
datetime
flask_sqlalchemy
pymysql
python-dotenv
flask-cors